requirements.txt
mysql-connector-python==8.2.0
requests==2.31.0
Flask==3.0.0

